test: close qa gaps on phase p1 ownership helpers
