+++ crates/atm-core/src/team_admin.rs
@@ -1,4 +1,3 @@
-use std::collections::BTreeSet;
 use std::fs;
 use std::path::{Path, PathBuf};
 
@@ -11,10 +10,12 @@ use crate::address::validate_path_segment;
 use crate::config::{load_config, load_team_config, resolve_team};
 use crate::error::{AtmError, AtmErrorCode, AtmErrorKind};
 use crate::home;
-use crate::mailbox::lock;
 use crate::persistence;
 use crate::schema::{AgentMember, TeamConfig};
 use crate::types::{AgentName, TeamName};
+
+#[path = "team_admin/restore.rs"]
+mod restore;
 
 /// One discovered team and its current member count.
 #[derive(Debug, Clone, Serialize, PartialEq, Eq)]
@@ -367,99 +368,7 @@ pub fn backup_team(request: BackupRequest) -> Result<BackupOutcome, AtmError> {
 /// config-last persistence fails. Failure to remove the restore marker after a
 /// successful restore is degraded to a warning-only follow-up path.
 pub fn restore_team(request: RestoreRequest) -> Result<RestoreResult, AtmError> {
-    let team_dir = home::team_dir_from_home(&request.home_dir, &request.team)?;
-    if !team_dir.exists() {
-        return Err(AtmError::team_not_found(&request.team));
-    }
-    let current_config = load_team_config(&team_dir)?;
-    let backup_dir = locate_backup_dir(&request.home_dir, &request.team, request.from.as_deref())?;
-    let backup_config = load_team_config(&backup_dir)?;
-
-    let members_to_restore = backup_config
-        .members
-        .iter()
-        .filter(|member| member.name != "team-lead")
-        .filter(|member| {
-            !current_config
-                .members
-                .iter()
-                .any(|existing| existing.name == member.name)
-        })
-        .map(|member| member.name.clone())
-        .collect::<Vec<_>>();
-    let members_to_restore_set = members_to_restore.iter().cloned().collect::<BTreeSet<_>>();
-
-    let mut inboxes_to_restore = list_backup_inboxes(&backup_dir)?;
-    inboxes_to_restore.retain(|name| {
-        if name == "team-lead.json" {
-            return false;
-        }
-        name.strip_suffix(".json")
-            .is_some_and(|member| members_to_restore_set.contains(member))
-    });
-    let tasks_to_restore = count_numeric_task_files(&backup_dir.join("tasks"))?;
-
-    if request.dry_run {
-        return Ok(RestoreResult::DryRun(RestorePlan {
-            action: "restore",
-            team: request.team.into(),
-            backup_path: backup_dir,
-            dry_run: true,
-            would_restore_members: members_to_restore.into_iter().map(Into::into).collect(),
-            would_restore_inboxes: inboxes_to_restore,
-            would_restore_tasks: tasks_to_restore,
-        }));
-    }
-
-    prepare_restore_workspace(&team_dir, &backup_dir)?;
-    let mut updated_config = current_config.clone();
-    for member in &backup_config.members {
-        if member.name == "team-lead" {
-            continue;
-        }
-        if updated_config
-            .members
-            .iter()
-            .any(|existing| existing.name == member.name)
-        {
-            continue;
-        }
-        let mut restored = member.clone();
-        clear_runtime_member_state(&mut restored);
-        updated_config.members.push(restored);
-    }
-    if let Some(value) = current_config.extra.get("leadSessionId") {
-        updated_config
-            .extra
-            .insert("leadSessionId".to_string(), value.clone());
-    }
-
-    apply_restored_inboxes(&team_dir, &backup_dir, &inboxes_to_restore)?;
-
-    let tasks_dir = tasks_dir_from_home(&request.home_dir, &request.team)?;
-    restore_task_state_from_backup(&backup_dir.join("tasks"), &tasks_dir)?;
-    write_team_config(&team_dir, &updated_config).map_err(|error| {
-        error.with_recovery("Check team config permissions and rerun `atm teams restore`.")
-    })?;
-    let marker_cleanup_error = clear_restore_marker(&team_dir).err();
-    cleanup_restore_workspace(&team_dir)?;
-    if let Some(error) = marker_cleanup_error {
-        warn!(
-            code = %AtmErrorCode::WarningRestoreInProgress,
-            %error,
-            team = %request.team,
-            "restore completed but the stale restore marker could not be removed"
-        );
-    }
-
-    Ok(RestoreResult::Applied(RestoreOutcome {
-        action: "restore",
-        team: request.team.into(),
-        backup_path: backup_dir,
-        members_restored: members_to_restore.len(),
-        inboxes_restored: inboxes_to_restore.len(),
-        tasks_restored: tasks_to_restore,
-    }))
+    restore::restore_team(request)
 }
 
 fn member_summary(member: &AgentMember) -> MemberSummary {
@@ -603,222 +512,6 @@ where
     Ok(())
 }
 
-fn locate_backup_dir(
-    home_dir: &Path,
-    team: &str,
-    explicit: Option<&Path>,
-) -> Result<PathBuf, AtmError> {
-    if let Some(path) = explicit {
-        if !path.is_dir() {
-            return Err(AtmError::missing_document(format!(
-                "backup directory not found: {}",
-                path.display()
-            )));
-        }
-        return Ok(path.to_path_buf());
-    }
-
-    let root = backup_root_from_home(home_dir, team)?;
-    if !root.exists() {
-        return Err(AtmError::missing_document(format!(
-            "no backup found for team '{}'",
-            team
-        )));
-    }
-    let mut entries = fs::read_dir(&root)
-        .map_err(|error| {
-            AtmError::file_policy(format!(
-                "failed to read backup directory {}: {error}",
-                root.display()
-            ))
-            .with_source(error)
-            .with_recovery("Check backup directory permissions or pass an explicit --from path.")
-        })?
-        .filter_map(Result::ok)
-        .map(|entry| entry.path())
-        .filter(|path| path.is_dir())
-        .collect::<Vec<_>>();
-    entries.sort();
-    entries
-        .pop()
-        .ok_or_else(|| AtmError::missing_document(format!("no backup found for team '{}'", team)))
-}
-
-fn list_backup_inboxes(backup_dir: &Path) -> Result<Vec<String>, AtmError> {
-    let inbox_dir = backup_dir.join("inboxes");
-    if !inbox_dir.exists() {
-        return Ok(Vec::new());
-    }
-
-    let mut names = fs::read_dir(&inbox_dir)
-        .map_err(|error| {
-            AtmError::mailbox_read(format!(
-                "failed to read backup inbox directory {}: {error}",
-                inbox_dir.display()
-            ))
-            .with_source(error)
-            .with_recovery("Check backup inbox permissions and retry the restore.")
-        })?
-        .filter_map(Result::ok)
-        .filter(|entry| entry.path().is_file())
-        .map(|entry| entry.file_name().to_string_lossy().to_string())
-        .collect::<Vec<_>>();
-    names.sort();
-    Ok(names)
-}
-
-fn count_numeric_task_files(tasks_dir: &Path) -> Result<usize, AtmError> {
-    if !tasks_dir.exists() {
-        return Ok(0);
-    }
-    let count = fs::read_dir(tasks_dir)
-        .map_err(|error| {
-            AtmError::file_policy(format!(
-                "failed to read task directory {}: {error}",
-                tasks_dir.display()
-            ))
-            .with_source(error)
-            .with_recovery("Check task directory permissions and retry the restore.")
-        })?
-        .filter_map(Result::ok)
-        .map(|entry| entry.path())
-        .filter(|path| {
-            path.is_file()
-                && path.extension().and_then(|ext| ext.to_str()) == Some("json")
-                && path
-                    .file_stem()
-                    .and_then(|stem| stem.to_str())
-                    .and_then(|stem| stem.parse::<u64>().ok())
-                    .is_some()
-        })
-        .count();
-    Ok(count)
-}
-
-fn restore_task_bucket(src: &Path, dst: &Path) -> Result<(), AtmError> {
-    if !src.exists() {
-        fs::create_dir_all(dst).map_err(|error| {
-            AtmError::file_policy(format!(
-                "failed to create task directory {}: {error}",
-                dst.display()
-            ))
-            .with_source(error)
-            .with_recovery("Check task directory permissions and rerun the restore.")
-        })?;
-        return Ok(());
-    }
-
-    let staging = dst.with_file_name(format!(
-        ".{}.restore.{}",
-        dst.file_name()
-            .and_then(|name| name.to_str())
-            .unwrap_or("tasks"),
-        std::process::id()
-    ));
-    if staging.exists() {
-        fs::remove_dir_all(&staging).map_err(|error| {
-            AtmError::file_policy(format!(
-                "failed to clear task staging directory {}: {error}",
-                staging.display()
-            ))
-            .with_source(error)
-            .with_recovery("Check task staging directory permissions and rerun the restore.")
-        })?;
-    }
-    copy_regular_files(src, &staging, |name| {
-        name == ".highwatermark" || name.ends_with(".json")
-    })?;
-
-    if dst.exists() {
-        fs::remove_dir_all(dst).map_err(|error| {
-            AtmError::file_policy(format!(
-                "failed to clear existing task directory {}: {error}",
-                dst.display()
-            ))
-            .with_source(error)
-            .with_recovery("Check task directory permissions and rerun the restore.")
-        })?;
-    }
-    if let Some(parent) = dst.parent() {
-        fs::create_dir_all(parent).map_err(|error| {
-            AtmError::file_policy(format!(
-                "failed to create task parent directory {}: {error}",
-                parent.display()
-            ))
-            .with_source(error)
-            .with_recovery("Check task parent directory permissions and rerun the restore.")
-        })?;
-    }
-    fs::rename(&staging, dst).map_err(|error| {
-        AtmError::file_policy(format!(
-            "failed to install restored task directory {}: {error}",
-            dst.display()
-        ))
-        .with_source(error)
-        .with_recovery("Check task directory permissions and rerun the restore.")
-    })?;
-    Ok(())
-}
-
-fn recompute_highwatermark(tasks_dir: &Path) -> Result<usize, AtmError> {
-    fs::create_dir_all(tasks_dir).map_err(|error| {
-        AtmError::file_policy(format!(
-            "failed to create task directory {}: {error}",
-            tasks_dir.display()
-        ))
-        .with_source(error)
-        .with_recovery("Check task directory permissions and rerun the restore.")
-    })?;
-
-    let max_id = fs::read_dir(tasks_dir)
-        .map_err(|error| {
-            AtmError::file_policy(format!(
-                "failed to read task directory {}: {error}",
-                tasks_dir.display()
-            ))
-            .with_source(error)
-            .with_recovery("Check task directory permissions and rerun the restore.")
-        })?
-        .filter_map(Result::ok)
-        .map(|entry| entry.path())
-        .filter(|path| {
-            path.is_file() && path.extension().and_then(|ext| ext.to_str()) == Some("json")
-        })
-        .filter_map(|path| {
-            path.file_stem()
-                .and_then(|stem| stem.to_str())
-                .and_then(|stem| stem.parse::<usize>().ok())
-        })
-        .max()
-        .unwrap_or(0);
-
-    persistence::atomic_write_string(
-        &tasks_dir.join(".highwatermark"),
-        &format!("{max_id}\n"),
-        AtmErrorKind::FilePolicy,
-        "task highwatermark",
-        "Check task directory permissions and rerun the restore.",
-    )?;
-    Ok(max_id)
-}
-
-fn clear_runtime_member_state(member: &mut AgentMember) {
-    member.tmux_pane_id.clear();
-    for key in [
-        "backendType",
-        "sessionId",
-        "activity",
-        "status",
-        "lastAliveAt",
-        "processId",
-        "isActive",
-        "lastActive",
-        "paneId",
-    ] {
-        member.extra.remove(key);
-    }
-}
-
 fn normalize_tmux_pane_id(pane_id: Option<&str>) -> Result<Option<String>, AtmError> {
     let Some(raw) = pane_id.map(str::trim).filter(|value| !value.is_empty()) else {
         return Ok(None);
@@ -838,171 +531,6 @@ fn normalize_tmux_pane_id(pane_id: Option<&str>) -> Result<Option<String>, AtmEr
     .with_recovery(
         "Pass `--pane-id $(tmux display-message -p '#{pane_id}')` or a bare numeric pane id when registering a tmux-backed member.",
     ))
-}
-
-fn restore_marker_path(team_dir: &Path) -> PathBuf {
-    team_dir.join(".restore-in-progress")
-}
-
-fn restore_staging_dir(team_dir: &Path) -> PathBuf {
-    team_dir.join(".restore-staging")
-}
-
-fn restore_staging_inboxes_dir(team_dir: &Path) -> PathBuf {
-    restore_staging_dir(team_dir).join("inboxes")
-}
-
-fn prepare_restore_staging_dir(team_dir: &Path) -> Result<(), AtmError> {
-    let staging_root = restore_staging_dir(team_dir);
-    if staging_root.exists() {
-        fs::remove_dir_all(&staging_root).map_err(|error| {
-            AtmError::file_policy(format!(
-                "failed to clear restore staging directory {}: {error}",
-                staging_root.display()
-            ))
-            .with_source(error)
-            .with_recovery("Check restore staging permissions and rerun `atm teams restore`.")
-        })?;
-    }
-    Ok(())
-}
-
-fn copy_restored_inbox_to_staging(from: &Path, staged: &Path) -> Result<u64, std::io::Error> {
-    if std::env::var_os("ATM_TEST_FAIL_RESTORE_INBOX_STAGE").is_some() {
-        return Err(std::io::Error::other(format!(
-            "forced inbox staging failure for {}",
-            staged.display()
-        )));
-    }
-    fs::copy(from, staged)
-}
-
-fn prepare_restore_workspace(team_dir: &Path, backup_dir: &Path) -> Result<(), AtmError> {
-    prepare_restore_staging_dir(team_dir)?;
-    write_restore_marker(team_dir, backup_dir)
-}
-
-fn cleanup_restore_workspace(team_dir: &Path) -> Result<(), AtmError> {
-    let staging_root = restore_staging_dir(team_dir);
-    if !staging_root.exists() {
-        return Ok(());
-    }
-
-    fs::remove_dir_all(&staging_root).map_err(|error| {
-        AtmError::file_policy(format!(
-            "failed to remove restore staging directory {}: {error}",
-            staging_root.display()
-        ))
-        .with_source(error)
-        .with_recovery("Remove the restore staging directory after confirming the restore completed successfully.")
-    })
-}
-
-fn apply_restored_inboxes(
-    team_dir: &Path,
-    backup_dir: &Path,
-    inboxes_to_restore: &[String],
-) -> Result<(), AtmError> {
-    let inboxes_dir = team_dir.join("inboxes");
-    fs::create_dir_all(&inboxes_dir).map_err(|error| {
-        AtmError::mailbox_write(format!(
-            "failed to create inbox directory {}: {error}",
-            inboxes_dir.display()
-        ))
-        .with_source(error)
-        .with_recovery("Check inbox directory permissions and rerun `atm teams restore`.")
-    })?;
-    lock::sweep_stale_lock_sentinels(&inboxes_dir).map_err(|error| {
-        error.with_recovery("Check inbox directory permissions and rerun `atm teams restore`.")
-    })?;
-
-    let inbox_staging_dir = restore_staging_inboxes_dir(team_dir);
-    fs::create_dir_all(&inbox_staging_dir).map_err(|error| {
-        AtmError::mailbox_write(format!(
-            "failed to create inbox restore staging directory {}: {error}",
-            inbox_staging_dir.display()
-        ))
-        .with_source(error)
-        .with_recovery("Check inbox staging permissions and rerun `atm teams restore`.")
-    })?;
-    for inbox_name in inboxes_to_restore {
-        let from = backup_dir.join("inboxes").join(inbox_name);
-        let staged = inbox_staging_dir.join(inbox_name);
-        copy_restored_inbox_to_staging(&from, &staged).map_err(|error| {
-            AtmError::mailbox_write(format!(
-                "failed to stage restored inbox {} from {}: {error}",
-                staged.display(),
-                from.display()
-            ))
-            .with_source(error)
-            .with_recovery("Check inbox permissions and backup integrity, then rerun the restore.")
-        })?;
-    }
-    for inbox_name in inboxes_to_restore {
-        let staged = inbox_staging_dir.join(inbox_name);
-        let to = inboxes_dir.join(inbox_name);
-        fs::rename(&staged, &to).map_err(|error| {
-            AtmError::mailbox_write(format!(
-                "failed to install restored inbox {} from {}: {error}",
-                to.display(),
-                staged.display()
-            ))
-            .with_source(error)
-            .with_recovery("Check inbox permissions and rerun `atm teams restore`.")
-        })?;
-    }
-    Ok(())
-}
-
-fn restore_task_state_from_backup(
-    backup_tasks_dir: &Path,
-    tasks_dir: &Path,
-) -> Result<usize, AtmError> {
-    restore_task_bucket(backup_tasks_dir, tasks_dir)?;
-    recompute_highwatermark(tasks_dir)
-}
-
-fn write_restore_marker(team_dir: &Path, backup_dir: &Path) -> Result<(), AtmError> {
-    let marker = restore_marker_path(team_dir);
-    let payload = serde_json::json!({
-        "backup_path": backup_dir,
-        "pid": std::process::id(),
-        "timestamp": Utc::now().to_rfc3339(),
-    });
-    let bytes = serde_json::to_vec_pretty(&payload).map_err(AtmError::from)?;
-    persistence::atomic_write_bytes(
-        &marker,
-        &bytes,
-        AtmErrorKind::FilePolicy,
-        "restore marker",
-        "Check team directory permissions and rerun `atm teams restore`.",
-    )
-}
-
-fn clear_restore_marker(team_dir: &Path) -> Result<(), AtmError> {
-    let marker = restore_marker_path(team_dir);
-    if !marker.exists() {
-        return Ok(());
-    }
-
-    if std::env::var_os("ATM_TEST_FAIL_RESTORE_MARKER_REMOVE").is_some() {
-        return Err(AtmError::file_policy(format!(
-            "failed to remove restore marker {}: forced test failure",
-            marker.display()
-        ))
-        .with_recovery(
-            "Remove the stale restore marker after verifying the restored team state.",
-        ));
-    }
-
-    fs::remove_file(&marker).map_err(|error| {
-        AtmError::file_policy(format!(
-            "failed to remove restore marker {}: {error}",
-            marker.display()
-        ))
-        .with_source(error)
-        .with_recovery("Remove the stale restore marker after verifying the restored team state.")
-    })
 }
 
 #[cfg(test)]
